requirements.txt
Flask
google-cloud-vision
google-genai
python-dotenv
python-docx
gunicorn